bug fixes + group_id
