+++ sentinel-scripts/edge_process.py
@@ -32,6 +32,8 @@ import pandas as pd
 from PIL import Image
 from PIL import ImageChops
 import logging
+import utils
+
 
 logger = logging.getLogger('cnn')
 
@@ -316,12 +318,20 @@ def main(alg,data_directory,quantize_type, algorithm_type = 'detection', batch =
     # Defining the minimum confidence required
     ai_sensitivity = alg['sensitivity'][0]
 
-    # Loading the model into tensorflow engine
-    if os.environ.get('version').startswith('0'):
-        interpreter = DetectionEngine(model)
-    if os.environ.get('version').startswith('1'):
-        interpreter = make_interpreter(model)
-        interpreter.allocate_tensors()
+    k = 0
+    while k < 3:
+        try:
+            # Loading the model into tensorflow engine
+            if os.environ.get('version').startswith('0'):
+                interpreter = DetectionEngine(model)
+            if os.environ.get('version').startswith('1'):
+                interpreter = make_interpreter(model)
+                interpreter.allocate_tensors()
+            break
+        except Exception as e:
+            logging.error('No TPU Found. Trying again')
+        k = k + 1
+    utils.shutdown(0)
 
     x = 0
     directories = [str(data_directory)]
@@ -343,7 +353,7 @@ def main(alg,data_directory,quantize_type, algorithm_type = 'detection', batch =
             #Rename the file
             os.rename('{}/{}'.format(directories[x],original_directory_list[i]),'{}/{}.{}'.format(directories[x],uuid.uuid1(),fileExtension))
             i = i+1
-        
+
         directory_list = os.listdir(directories[x])
         # initializing variables to enable grouping of files
         previous_confidence = 0
@@ -360,7 +370,7 @@ def main(alg,data_directory,quantize_type, algorithm_type = 'detection', batch =
             # logger.info("HERE WE GO")
             # logger.info('{}/{}'.format(directories[x],directory_list[k]))
 
-            if k > 0: 
+            if k > 0:
                 timeFile = int(os.path.getmtime('{}/{}'.format(directories[x], file)))
                 timeFileBefore = int(os.path.getmtime('{}/{}'.format(directories[x],directory_list[k-1])))
             else:
@@ -369,17 +379,17 @@ def main(alg,data_directory,quantize_type, algorithm_type = 'detection', batch =
             logger.info('timeFile - timeFileBefore')
             logger.info(timeFile - timeFileBefore)
 
-            #If it is the first photo in the directory, we can assume it is a new group 
+            #If it is the first photo in the directory, we can assume it is a new group
             #If it is the first ever photo in the csv, we set the key to 1
             if k == 0:
                 try:
-                    group_key = alg_df['group_id'].iloc[-1]+1 
+                    group_key = alg_df['group_id'].iloc[-1]+1
                     logger.info(group_key)
                 except Exception as e:
                     group_key = 1
             #If the gap between photos after the first is smaller than 30 seconds, it is the same group
             elif (timeFile - timeFileBefore) < 30:
-                group_key = alg_df['group_id'].iloc[-1] 
+                group_key = alg_df['group_id'].iloc[-1]
             #If the gap is larger than 30 seconds, then we define a new group key
             else:
                 group_key = alg_df['group_id'].iloc[-1] + 1
@@ -395,6 +405,8 @@ def main(alg,data_directory,quantize_type, algorithm_type = 'detection', batch =
                 # Checking if number of files checked has exceeded the "batch" variable (in production this should be inf)
                 if k == batch:
                     break
+
+
             ## Check that the file is actually a processable photo
             #  Feature: Add ability to process videos
             if file.endswith(".jpeg") or file.endswith(".JPG") or file.endswith(".jpg") or file.endswith(".JPEG") or file.endswith(".png") or file.endswith(".PNG"):
@@ -408,26 +420,49 @@ def main(alg,data_directory,quantize_type, algorithm_type = 'detection', batch =
                         alg_df = alg_df.append(meta_df,ignore_index=True)
                         tempalg_df=alg_df
                         logger.info(alg_df)
-
-                        # Adding group confidence from linked confidence between inferences
-                        m = 0
-                        while m < len(meta_df):
-                            # if the previous data inference within the same group had the same class, add the confidences.
-                            # note that this will only run if all detections within the current image are also the same as the previous image
-                            if previous_class == meta_df['class'][m]:
-                                try:
-                                    previous_confidence = meta_df['confidence'][m] + previous_confidence
-                                except Exception as e:
-                                    previous_confidence = previous_confidence
-                            previous_class = meta_df['class'][m]
-                            m = m+1
                     else:
                         logger.error('Type of algorithm not yet supported')
             else:
+                logger.error('File Format not yet supported')
                 break
-            # Adding the group confidence to any data point that has the same group key
-            alg_df.loc[alg_df['group_id'] == group_key,'group_confidence'] = previous_confidence
-                    
+
+
+            ### Adding the group confidence to any data point that has the same group key
+
+            # Finding all group id's and putting them in list
+            group_keys = alg_df.group_id.unique()
+
+            # Loop through each group id. (Should do a for loop, but I get confused by those things)
+            y = 0
+            while y < len(group_keys):
+
+                # Segment the group we are working with
+                group = alg_df.loc[alg_df['group_id'] == group_keys[y]]
+
+                # Confidence algorithm
+                """
+                Variables
+                - Number of possible classes
+                - Number of events
+                - Confidence of individual events
+                - Number of unique classes within the group
+
+                v1: Hyperbolic towards 1, with reset if different class
+                """
+
+
+                group_confidence = 0
+                m = 0
+                class_id = group['class_id'][0]
+                while m = len(group):
+                    if class_id == group['class_id'][m]:
+                        group_confidence = (1-group_confidence)*group['confidence'][m]
+                    else:
+                        group_confidence = 0
+                    class_id = group['class_id'][m]
+                    m = m + 1
+                alg_df.loc[alg_df['group_id'] == group_keys[y],'group_confidence'] = group_confidence
+                y = y + 1
             # Moving on to next file
             k = k + 1
         x = x + 1
